- systemSCs.go code split
